feature/#73-ManagePageRouterUpdate
Manage page router update.
