+++ src/consts/userType.js
@@ -11,3 +11,6 @@ export const SET_TOKEN = 'setToken';
 export const LOGIN_URL =  process.env.VUE_APP_API_URL + '/users/sign-in';
 export const SIGNUP_URL =  process.env.VUE_APP_API_URL + '/users/sign-up';
 export const GET_ME_INFO =  process.env.VUE_APP_API_URL + '/apis/me';
+
+export const ALL_USER_URL = process.env.VUE_APP_API_URL_TEST + '/users';
+export const ACTIVE_USER_URL = process.env.VUE_APP_API_URL_TEST + '/users';
